gcp: avoid using node's metadata service creds
Unlike Azure and AWS, GCP requires `GOOGLE_APPLICATION_CREDENTIALS` to
point to a file instead of using raw environment variables.

This was working before because GCP instances uses the metadata service
that lets workloads authenticate using the worker node’s service
account. However the node's OAuth scopes are too restrictive, blocking
some API calls. I only caught this while working on the next commit.

A better fix could be to mount `peer-pods-secret` as a file in the
DaemonSet, but to avoid unnecessary deployment changes right now, this
writes the secret to a file and sets `GOOGLE_APPLICATION_CREDENTIALS` so
authentication works as expected.

Signed-off-by: Beraldo Leal <bleal@redhat.com>